backend/pyproject.toml
[project]
name = "saju-backend"
version = "0.1.0"
description = "Saju Hub FastAPI backend"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.32.1",
  "pydantic==2.10.3",
  "google-genai==1.11.0",
  "httpx==0.28.1",
  "sqlalchemy==2.0.36",
  "psycopg2-binary==2.9.10",
]

[dependency-groups]
dev = [
  "pytest==8.3.4",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
